Move LollipopRenderer into lollipop package
