+++ packages/jbrowse-web/package.json
@@ -22,6 +22,7 @@
     "@gmod/jbrowse-plugin-data-management": "^2.0.0",
     "@gmod/jbrowse-plugin-jbrowse1": "^2.0.0",
     "@gmod/jbrowse-plugin-linear-genome-view": "^2.0.0",
+    "@gmod/jbrowse-plugin-lollipop": "^2.0.0",
     "@gmod/jbrowse-plugin-sequence": "^2.0.0",
     "@gmod/jbrowse-plugin-trackhub-registry": "^2.0.0",
     "@gmod/jbrowse-plugin-wiggle": "^2.0.0",
