+++ packages/lollipop/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@gmod/jbrowse-plugin-lollipop",
+  "version": "2.0.0",
+  "description": "JBrowse 2 plugin for lollipops",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "cd ../..; jest packages/lollipop"
+  },
+  "author": "Garrett Stevens",
+  "license": "MIT",
+  "dependencies": {
+    "@gmod/jbrowse-core": "^2.0.0",
+    "@material-ui/core": "3.9.3",
+    "mobx-react": "^5.3.6",
+    "prop-types": "^15.6.2",
+    "react": "^16.8.6"
+  }
+}
