+++ packages/lollipop/src/LollipopRenderer/components/LollipopRendering.js
@@ -1,12 +1,10 @@
-import React, { Component } from 'react'
+import { readConfObject } from '@gmod/jbrowse-core/configuration'
+import { PropTypes as CommonPropTypes } from '@gmod/jbrowse-core/mst-types'
+import { bpToPx } from '@gmod/jbrowse-core/util'
+import { withStyles } from '@material-ui/core'
 import { observer } from 'mobx-react'
 import ReactPropTypes from 'prop-types'
-
-import { withStyles } from '@material-ui/core'
-import { PropTypes as CommonPropTypes } from '@gmod/jbrowse-core/mst-types'
-
-import { bpToPx } from '@gmod/jbrowse-core/util'
-import { readConfObject } from '@gmod/jbrowse-core/configuration'
+import React, { Component } from 'react'
 import Lollipop from './Lollipop'
 import Stick from './Stick'
