+++ packages/lollipop/src/LollipopRenderer/configSchema.js
