+++ packages/lollipop/src/LollipopRenderer/index.js
@@ -1,0 +1,3 @@
+export { default as ReactComponent } from './components/LollipopRendering'
+export { default as configSchema } from './configSchema'
+export { default } from './LollipopRenderer'
